Move core/modules datastore files
