+++ assets/js/googlesitekit/modules/datastore/index.js
