+++ assets/js/googlesitekit/modules/datastore/index.test.js
