+++ assets/js/googlesitekit/modules/index.js
@@ -25,7 +25,7 @@ import Data from 'googlesitekit-data';
 import { createModuleStore } from './create-module-store';
 // This import has a side-effect: it automatically registers the "core/modules"
 // store on `googlesitekit.data`.
-import { STORE_NAME } from '../datastore/modules';
+import { STORE_NAME } from '../modules/datastore';
 
 const Modules = {
 	createModuleStore,
